requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
black==22.1.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.3
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
pyparsing==3.0.7
pytz==2022.1
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.0
zope.event==4.5.0
zope.interface==5.4.0
